Refactored Services and Controllers

+++ src/main/java/com/miniconomy/commercial_bank_service/financial_management/response/AccountResponse.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class AccountResponse
 {
   private String accountName;
-  private Double accountBalance;  
+  private Integer accountBalance;  
 }
